requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
anyio==4.13.0
APScheduler==3.11.2
asyncpg==0.31.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
ddgs==9.12.0
docker==7.1.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
playwright==1.58.0
primp==1.2.0
propcache==0.4.1
pyee==13.0.1
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.33.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.23.0
